sessions: move ISessionsPartService and ISessionsViewService to services layer
Move ISessionsPartService and ISessionsViewService (+ impl) from browser layer to services/sessions/browser layer to make them accessible to lower-layer services.

This enables lower-layer services like sessionNavigation and sessionsListModelService to depend on these view services directly, which is a prerequisite for removing the activeSession mirror from ISessionsManagementService.

- Moved: browser/parts/sessionsPartService.ts → services/sessions/browser/sessionsPartService.ts
- Moved: browser/sessionsViewService.ts → services/sessions/browser/sessionsViewService.ts
- Updated all ~30 import sites across production code, tests, and entry points
