+++ src/vs/sessions/browser/parts/sessionDropTarget.ts
@@ -15,7 +15,7 @@ import { IThemeService, Themable } from '../../../platform/theme/common/themeSer
 import { EDITOR_DRAG_AND_DROP_BACKGROUND } from '../../../workbench/common/theme.js';
 import { DraggedSessionIdentifier } from '../dnd.js';
 import { ISessionsManagementService } from '../../services/sessions/common/sessionsManagement.js';
-import { ISessionsViewService } from '../sessionsViewService.js';
+import { ISessionsViewService } from '../../services/sessions/browser/sessionsViewService.js';
 
 /** Side of a target view where a dragged session can be dropped. */
 type DropSide = 'left' | 'right';
