+++ src/vs/sessions/sessions.common.main.ts
@@ -448,7 +448,7 @@ import '../workbench/contrib/opener/browser/opener.contribution.js';
 import './browser/paneCompositePartService.js';
 import './browser/parts/editorParts.js';
 import './browser/parts/sessionsParts.js';
-import './browser/sessionsViewService.js';
+import './services/sessions/browser/sessionsViewService.js';
 import './browser/parts/menubar.contribution.js';
 import './browser/layoutActions.js';
